Define shared execution intent contract for order router
